+++ src/IContext.ts
@@ -1,0 +1,4 @@
+export default interface IContext {
+    expand(uri: string): string;
+    shorten(uri: string): string;
+}