+++ CHANGELOG.md
@@ -7,6 +7,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Update `rdf-ext` to version 2.5.1.
+- Update `@types/rdf-ext` to version 2.5.0.
+- Do not require `semanticType` in constructor.
+
+### Removed
+
+- The `init` method in `SemanticObject` and `SemanticObjectAnonymous`.
+
+### Added
+
+- New `Semantizer` and `Context` classes with corresponding interfaces.
+- Ability to provide a context and pass prefixed URIs.
+
 ## [1.0.0-alpha.2] - 2023-05-23
 
 ### Added
@@ -34,7 +49,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release.
 
-[unreleased]: https://github.com/assemblee-virtuelle/semantizer-typescript/compare/v1.0.0-alpha.1...HEAD
+[unreleased]: https://github.com/assemblee-virtuelle/semantizer-typescript/compare/v1.0.0-alpha.2...HEAD
 [1.0.0-alpha.2]: https://github.com/assemblee-virtuelle/semantizer-typescript/compare/v1.0.0-alpha.1...v1.0.0-alpha.2
 [1.0.0-alpha.1]: https://github.com/assemblee-virtuelle/semantizer-typescript/compare/v1.0.0-alpha...v1.0.0-alpha.1
 [1.0.0-alpha]: https://github.com/assemblee-virtuelle/semantizer-typescript/releases/tag/v1.0.0-alpha